requirements.txt
# Minimal working requirements for Azure deployment
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Essential dependencies 
pydantic==2.8.2
pydantic-core==2.20.1
typing-extensions>=4.11,<5
python-dotenv==1.0.0
requests==2.31.0
setuptools==70.0.0

# LangGraph for agent orchestration
langgraph==0.2.74
langchain-core==0.3.29
langchain-openai==0.2.14

# OpenAI
openai>=1.58.1

# Azure services
azure-storage-blob==12.19.0
azure-servicebus==7.11.4

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5

# Email processing
email-validator==2.1.0
beautifulsoup4==4.12.3

# Other essentials
python-multipart==0.0.6
structlog==23.2.0
aiohttp==3.9.1
tenacity==8.2.3

# Azure AI Search and Analytics
azure-search-documents==11.4.0
applicationinsights==0.11.10
numpy==1.24.3
scipy==1.11.4

# WebSocket and streaming support
websockets==12.0
sse-starlette==1.8.2
pyjwt==2.8.0

# Microsoft Teams Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botframework-connector==4.15.0

# Redis caching for GPT-5-mini optimization
redis==5.0.1

# Azure monitoring and security
azure-monitor-opentelemetry==1.2.0
azure-monitor-query==1.2.0
azure-keyvault-secrets==4.7.0
azure-keyvault-keys==4.8.0
azure-identity==1.15.0
azure-core==1.29.6

# Security and rate limiting
slowapi==0.1.9
cryptography==41.0.7

# Token counting for cost calculation
tiktoken>=0.7,<1

# System monitoring
psutil==5.9.6

# User agent parsing for analytics
user-agents==2.2.0

# Template engine for manifest generation
jinja2>=3.1.2

# Azure Front Door SDK for CDN management
azure-mgmt-frontdoor==1.1.0
azure-mgmt-cdn==13.1.1

# TalentWell email delivery
azure-communication-email==1.0.0
sendgrid==6.11.0

# TalentWell CSV processing
pandas==2.0.3
httpx==0.25.2
chardet==5.2.0
openpyxl==3.1.2
xlrd==2.0.1
aiofiles==23.2.1

# Scheduled job management
croniter==1.4.1

# Firecrawl v2 SDK for web scraping and enrichment
firecrawl-py==4.3.6
